Separate backend from frontend into Roll-Call Backend project
- Convert Roll-Call Backend from console app to ASP.NET Core Web API
- Move Models (15 classes), Services (Auth + DataTable), and DataTables (11 JSON files) from PF2EGM to Roll-Call Backend
- Add AuthController (sign-in, sign-up, sign-out, OAuth, status) and DataController (all game data endpoints)
- Add CORS support for the Blazor frontend origin
- Replace Supabase direct dependency in PF2EGM with HttpClient-based ApiAuthService
- Update Login and Register pages to inject IApiAuthService instead of ISupabaseAuthService
