+++ PF2EGM/Components/Pages/Login.razor
@@ -5,7 +5,7 @@
 @using LumexUI
 @using LumexUI.Common
 @using LumexUI.Shared.Icons
-@inject ISupabaseAuthService AuthService
+@inject IApiAuthService AuthService
 @inject NavigationManager Nav
 
 <PageTitle>Sign In</PageTitle>
