+++ PF2EGM/Components/Pages/Register.razor
@@ -3,7 +3,7 @@
 @using System.ComponentModel.DataAnnotations
 @using PF2EGM.Services
 @using LumexUI
-@inject ISupabaseAuthService AuthService
+@inject IApiAuthService AuthService
 @inject NavigationManager Nav
 
 <PageTitle>Create Account</PageTitle>
