--- PF2EGM/Models/Ability.cs
@@ -1,10 +1,0 @@
-﻿namespace PF2EGM.Models;
-
-public class Ability
-{
-    public string Name { get; set; }
-    
-    public string Description { get; set; }
-    
-    public List<Trait> Traits { get; set; }
-}
