--- PF2EGM/Models/Attribute.cs
@@ -1,12 +1,0 @@
-﻿namespace PF2EGM.Models;
-
-public class Attribute
-{
-    public string Name { get; set; }
-    
-    public string Abbreviation { get; set; }
-    
-    public int Score { get; set; }
-    
-    public List<Skill> Skills { get; set; }
-}
