--- PF2EGM/Models/Heritage.cs
@@ -1,9 +1,0 @@
-﻿namespace PF2EGM.Models;
-
-public class Heritage
-{
-    public string Name { get; set; }
-    public string Description { get; set; }
-    public Ancestry Ancestry { get; set; }
-    public List<Feat> FeatsRequiringHeritage { get; set; }
-}
