+++ PF2EGM/PF2EGM.csproj
@@ -9,7 +9,6 @@
 
     <ItemGroup>
       <PackageReference Include="LumexUI" Version="2.3.4" />
-      <PackageReference Include="Supabase" Version="1.1.1" />
       <PackageReference Include="TailwindMerge.NET" Version="1.3.0" />
     </ItemGroup>
