+++ Roll-Call Backend/Models/Attribute.cs
@@ -1,0 +1,9 @@
+namespace RollCallBackend.Models;
+
+public class Attribute
+{
+    public string Name { get; set; } = "";
+    public string Abbreviation { get; set; } = "";
+    public int Score { get; set; }
+    public List<Skill> Skills { get; set; } = [];
+}
