+++ Roll-Call Backend/Models/Heritage.cs
@@ -1,0 +1,9 @@
+namespace RollCallBackend.Models;
+
+public class Heritage
+{
+    public string Name { get; set; } = "";
+    public string Description { get; set; } = "";
+    public Ancestry? Ancestry { get; set; }
+    public List<Feat> FeatsRequiringHeritage { get; set; } = [];
+}
