+++ Roll-Call Backend/Services/SupabaseAuthService.cs
@@ -1,39 +1,38 @@
 using Supabase;
 using Supabase.Gotrue;
 using Supabase.Gotrue.Exceptions;
- 
-namespace PF2EGM.Services;
+
+namespace RollCallBackend.Services;
 
 public interface ISupabaseAuthService
 {
     Task<AuthResult> SignInAsync(string email, string password);
     Task<AuthResult> SignUpAsync(string email, string password);
     Task SignOutAsync();
-    Task<string> GetOAuthUrlAsync(string provider);
+    Task<string> GetOAuthUrlAsync(string provider, string redirectTo);
     Task<bool> IsAuthenticatedAsync();
 }
- 
+
 public record AuthResult(bool IsSuccess, string? ErrorMessage = null);
- 
+
 public class SupabaseAuthService : ISupabaseAuthService
 {
     private readonly Supabase.Client _client;
     private readonly ILogger<SupabaseAuthService> _logger;
- 
+
     public SupabaseAuthService(Supabase.Client client, ILogger<SupabaseAuthService> logger)
     {
         _client = client;
         _logger  = logger;
     }
 
-    //get sign in data
     public async Task<AuthResult> SignInAsync(string email, string password)
     {
         try
         {
             var session = await _client.Auth.SignIn(email, password);
             return session?.User is not null
-                ? new AuthResult(true) //authenticate sign in data from database
+                ? new AuthResult(true)
                 : new AuthResult(false, "Sign-in failed. Please check your credentials.");
         }
         catch (GotrueException ex)
@@ -48,14 +47,13 @@ public class SupabaseAuthService : ISupabaseAuthService
         }
     }
 
-    //get sign up data
     public async Task<AuthResult> SignUpAsync(string email, string password)
     {
         try
         {
             var session = await _client.Auth.SignUp(email, password);
             return session?.User is not null
-                ? new AuthResult(true) //create new sign in data in database
+                ? new AuthResult(true)
                 : new AuthResult(false, "Sign-up failed.");
         }
         catch (GotrueException ex)
@@ -63,25 +61,22 @@ public class SupabaseAuthService : ISupabaseAuthService
             return new AuthResult(false, FriendlyError(ex.Message));
         }
     }
- 
+
     public async Task SignOutAsync() => await _client.Auth.SignOut();
 
-    public async Task<string> GetOAuthUrlAsync(string provider)
+    public async Task<string> GetOAuthUrlAsync(string provider, string redirectTo)
     {
         var p = Enum.Parse<Supabase.Gotrue.Constants.Provider>(provider, ignoreCase: true);
-        var state = await _client.Auth.SignIn(p, new SignInOptions
-        {
-            RedirectTo = "https://localhost:5144/auth/callback"
-        });
+        var state = await _client.Auth.SignIn(p, new SignInOptions { RedirectTo = redirectTo });
         return state!.Uri!.ToString();
     }
- 
+
     public async Task<bool> IsAuthenticatedAsync()
     {
         var session = _client.Auth.CurrentSession;
         return session?.User is not null && session.ExpiresAt() > DateTime.UtcNow;
     }
- 
+
     private static string FriendlyError(string raw) => raw.ToLower() switch
     {
         var s when s.Contains("invalid login") => "Invalid email or password.",
