+++ Roll-Call Backend/appsettings.Development.json
@@ -4,7 +4,5 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
-  },
-  "AllowedHosts": "*",
-  "BackendUrl": "https://localhost:7001"
+  }
 }
