+++ Roll-Call Backend/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "Supabase": {
+    "Url":     "https://qwovodapgjzacewgephz.supabase.co",
+    "AnonKey": "sb_publishable_Bv3nGhYN8NgQHqy0SLxLAQ_TUDN0nsL"
+  },
+  "AllowedHosts": "*",
+  "AllowedOrigins": [ "https://localhost:7040", "http://localhost:5144" ]
+}
